requirements-core.txt
# Core Flask dependencies (lightweight)
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Database
pymongo==4.6.1

# Authentication
bcrypt==4.1.2
PyJWT==2.8.0

# Image processing
Pillow==10.0.1

# Data validation
pydantic==2.11.7

# Translation
deep-translator==1.11.4

# Cloud storage
cloudinary==1.44.1

# Production server
gunicorn==21.2.0
